Fixed the double to int error

+++ lib/Classes/modalSheet.dart
@@ -49,9 +49,8 @@ class _NewTransactionState extends State<NewTransaction> {
       );
       return;
     } else {
-      widget.newTransaction(
-        _expirationDateController, 'a', 'b', double.parse(quantityController.text)
-      );
+      widget.newTransaction(_expirationDateController, 'a', 'b',
+          int.parse(quantityController.text));
 
       Navigator.of(context).pop();
     }
